refactor repoCard to own component
