+++ packages/api/src/graphql/typeDefs/project.graphql
@@ -24,7 +24,12 @@ type Query {
 }
 
 type Mutation {
-  createProject(name: String!, description: String, type: ProjectType!, repositoryId: Int): Project
+  createProject(
+    name: String!
+    description: String
+    type: ProjectType!
+    repositoryId: Int
+  ): Project
   updateProject(
     id: ID!
     name: String
