+++ packages/client/src/components/Profile.tsx
@@ -18,8 +18,8 @@ export const Profile = () => {
     const settings = searchParams.get("settings");
     return settings ? settings : "user";
   });
-  const [dialogOpen, setDialogOpen] = useState<boolean>(false)
-  const [repositoryId, setRepositoryId] = useState<number | null>(null)
+  const [dialogOpen, setDialogOpen] = useState<boolean>(false);
+  const [repositoryId, setRepositoryId] = useState<number | null>(null);
 
   const handleSelectedSettings = (selectedItem: string) => {
     setSelectedSettings(selectedItem);
@@ -167,7 +167,13 @@ export const Profile = () => {
                   )} */}
                   {selectedSettings === "github" && (
                     <div className="mt-2 h-full overflow-y-auto no-scrollbar">
-                      {user && <Github user={user} setDialogOpen={setDialogOpen} setRepositoryId={setRepositoryId}/>}
+                      {user && (
+                        <Github
+                          user={user}
+                          setDialogOpen={setDialogOpen}
+                          setRepositoryId={setRepositoryId}
+                        />
+                      )}
                     </div>
                   )}
                 </div>
@@ -185,7 +191,7 @@ export const Profile = () => {
           className="absolute bottom-1/4 right-1/4 w-48 sm:w-64 h-48 sm:h-64 bg-purple-600/20 rounded-full blur-3xl translate-x-1/2 -z-10"
           aria-hidden="true"
         />
-        <CreateProjectDialog 
+        <CreateProjectDialog
           open={dialogOpen}
           setOpen={setDialogOpen}
           repositoryId={repositoryId}
